requirements.txt
Django==6.0.*
psycopg==3.1.20
python-decouple==3.8
argon2-cffi==23.1.*
celery[redis]==5.4.*
django-widget-tweaks==1.5.*